Title: 알파벳과 사칙연산, Time: -1ms, Memory: -1MB, Status: Compile Error - CodeTree

+++ 240419/알파벳과 사칙연산/calculations-with-alphabet.java
@@ -33,7 +33,7 @@ public class Main {
     public static void renewAns() {
         int idx = 1;
         
-        int temp = alp[str.charAt(0) - 'a'];
+        long temp = alp[str.charAt(0) - 'a'];
 
         while (idx < str.length()) {
             char cmd = str.charAt(idx);
